refactor: extract harden_flatpak funcs to utils
